+++ neo/VM/tests/test_execution_engine.py
@@ -3,6 +3,10 @@ from neo.VM.InteropService import *
 from neo.VM.ExecutionEngine import ExecutionEngine
 from neo.VM.ExecutionEngine import ExecutionContext
 from neo.VM import OpCode
+from neocore.Cryptography.Crypto import Crypto
+from mock import patch
+import binascii
+from neo.Core.TX.Transaction import ContractTransaction
 
 
 class VMTestCase(TestCase):
@@ -12,7 +16,7 @@ class VMTestCase(TestCase):
 
     def setUp(self):
 
-        self.engine = ExecutionEngine()
+        self.engine = ExecutionEngine(crypto=Crypto.Default())
         self.econtext = ExecutionContext()
 
     def test_add_operations(self):
@@ -36,3 +40,39 @@ class VMTestCase(TestCase):
         self.assertEqual(len(self.engine.EvaluationStack.Items), 1)
 
         self.assertEqual(self.engine.EvaluationStack.Items[0], StackItem.New(-1))
+
+    def test_verify_sig(self):
+
+        # push message ( should be hexlified )
+        self.engine.EvaluationStack.PushT(StackItem.New(b'616263646566'))
+
+        # sig
+        sig = binascii.unhexlify(b'cd0ca967d11cea78e25ad16f15dbe77672258bfec59ff3617c95e317acff063a48d35f71aa5ce7d735977412186e1572507d0f4d204c5bcb6c90e03b8b857fbd')
+        self.engine.EvaluationStack.PushT(StackItem.New(sig))
+
+        # pubkey
+        pubkey = binascii.unhexlify(b'036fbcb5e138c1ce5360e861674c03228af735a9114a5b7fb4121b8350129f3ffe')
+        self.engine.EvaluationStack.PushT(pubkey)
+
+        self.engine.ExecuteOp(OpCode.VERIFY, self.econtext)
+
+        res = self.engine.EvaluationStack.Pop()
+        self.assertEqual(res, StackItem.New(True))
+
+    def test_verify_sig_fail(self):
+
+        # push message ( should be hexlified )
+        self.engine.EvaluationStack.PushT(StackItem.New(b'616263646566'))
+
+        # sig
+        sig = binascii.unhexlify(b'cd0ca967d11cea78e25ad16f15dbe77672258bfec59ff3617c95e317acff063a48d35f71aa5ce7d735977412186e1572507d0f4d204c5bcb6c90e03b8b857fbd')
+        self.engine.EvaluationStack.PushT(StackItem.New(sig))
+
+        # pubkey
+        pubkey = binascii.unhexlify(b'036fbcb5e138c1ce5360e861674c03228af735a9114a5b7fb4121b8350129f3ffd')
+        self.engine.EvaluationStack.PushT(pubkey)
+
+        self.engine.ExecuteOp(OpCode.VERIFY, self.econtext)
+
+        res = self.engine.EvaluationStack.Pop()
+        self.assertEqual(res, StackItem.New(False))